third_party/blink/web_tests/external/wpt/merchant-validation/onmerchantvalidation-attribute.https-expected.txt

This is a testharness.js-based test.
[FAIL] Must have a onmerchantvalidation IDL attribute
  assert_idl_attribute: property "onmerchantvalidation" not found in prototype chain
[FAIL] onmerchantvalidation attribute is a generic handler for "merchantvalidation"
  assert_true: event did not fire expected true got false
[FAIL] onmerchantvalidation attribute is a handler for MerchantValidationEvent
  MerchantValidationEvent is not defined
[FAIL] onmerchantvalidation attribute and listeners both work
  MerchantValidationEvent is not defined
Harness: the test ran to completion.
